### 2023-07-06 #5624 @editorialbot
[PRE REVIEW]: CMakePPLang: An object-oriented extension to CMake
**Submitting author:** @ryanmrichard (<a href="http://orcid.org/0000-0003-4235-5179">Ryan M. Richard</a>)
**Repository:** https://github.com/CMakePP/CMakePPLang/
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
