### 2023-07-19 #5624 @ryanmrichard
@diehlpk, the last time we submitted a CMake paper to JOSS `robertodr` and `peanutfun` reviewed it.
