### 2023-07-24 8b0bd3d
Update README.rst (#97)
This PR is mostly just to trigger the documentation workflow. I changed around one line length in the README.
