+++ README.rst
@@ -49,7 +49,7 @@ CMakePPLang can be included in a CMake project by adding:
 
 to your ``CMakeLists.txt`` file. CMakePPLang will then be downloaded as part
 of the CMake configuration step. More detailed instructions, can be found
-`here <https://cmakepp.github.io/CMakePPLang/getting_started/obtaining_cmakepplanglanglang.html>`__.
+`here <https://cmakepp.github.io/CMakePPLang/getting_started/obtaining_cmakepplang.html>`__.
 
 *************
 Example Usage
