### 2023-07-27 b5541e3
Treat pointer types like desc
