### 2023-07-28 c605aac
Make pointers invariant, add tests for is_type, update some docs
