+++ cmake/cmakepp_lang/types/implicitly_convertible.cmake
@@ -108,6 +108,7 @@ function(cpp_implicitly_convertible _ic_result _ic_from _ic_to)
         if(_ic_to STREQUAL "type")
             set("${_ic_result}" TRUE PARENT_SCOPE)
         endif()
+    # Check if we're converting desc to pointer type or from pointer to desc
     elseif((_ic_from_original STREQUAL "desc" AND _ic_to_original MATCHES ".+[*]") OR (_ic_from_original MATCHES ".+[*]" AND _ic_to_original STREQUAL "desc"))
         set("${_ic_result}" TRUE PARENT_SCOPE)
     endif()
