### 2023-08-04 b74a371
Update docs/source/features/types.rst
Co-authored-by: Zachery Crandall <zachcran@gmail.com>
