### 2023-08-04 8b0afc3
Strip pointer asterisks from types in is_type instead of returning true
