### 2023-08-04 ef2b3a5
Add more tests for double pointers and invalid pointer types
