### 2023-08-07 b520db4
Add blurb about pointers in type relations summary
