### 2023-08-07 c540904
Add more pointer conversion tests
