+++ tests/types/implicitly_convertible.cmake
@@ -335,4 +335,16 @@ function("${test_cpp_implicitly_convertible}")
         cpp_implicitly_convertible(result bool* bool)
         ct_assert_equal(result FALSE)
     endfunction()
+
+    ct_add_section(NAME "double_pointer_to_pointer")
+    function("${double_pointer_to_pointer}")
+        cpp_implicitly_convertible(result bool** bool*)
+        ct_assert_equal(result FALSE)
+    endfunction()
+
+    ct_add_section(NAME "bool_to_pointer")
+    function("${bool_to_pointer}")
+        cpp_implicitly_convertible(result bool bool*)
+        ct_assert_equal(result FALSE)
+    endfunction()
 endfunction()
