### 2023-08-07 2ae9ba8
Update docs/source/features/types.rst
Co-authored-by: Zachery Crandall <zachcran@gmail.com>
