### 2023-08-07 4427a52
Treat pointer types like desc (#105)
* Treat pointer types like desc

* Make pointers invariant, add tests for is_type, update some docs

* Update documentation

* Update docs/source/features/types.rst

Co-authored-by: Zachery Crandall <zachcran@gmail.com>

* Strip pointer asterisks from types in is_type instead of returning true

* Add documentation about pointers to pointers

* Add more tests for double pointers and invalid pointer types

* Add blurb about pointers in type relations summary

* Add more pointer conversion tests

* Update docs/source/features/types.rst

Co-authored-by: Zachery
[…]
