### 2023-08-09 d3d3e68
Add tests for extension functions and constructors
