### 2023-08-09 6326b6e
Add underscore as permissible pointer specifier
