+++ tests/class/class.cmake
@@ -164,6 +164,11 @@ function(${_test_member})
         function("${fxn_c}" self a b)
             message("a = ${a}, b = ${b}")
         endfunction()
+
+        cpp_member(fxn_d MyClass bool*)
+        function("${fxn_d}" self a)
+            message("a = ${a}, a dereferenced = ${${a}}")
+        endfunction()
     cpp_end_class()
 
     ct_add_section(NAME "_member_single_param")
@@ -185,6 +190,15 @@ function(${_test_member})
         MyClass(CTOR my_instance)
         MyClass(fxn_c "${my_instance}" 1 "This is a string.")
         ct_assert_prints("a = 1, b = This is a string.")
+    endfunction()
+
+    ct_add_section(NAME "_member_bool_ptr_param")
+    function(${_member_bool_ptr_param})
+        MyClass(CTOR my_instance)
+        set(bool_var "TRUE")
+        set(bool_var_ptr bool_var)
+        MyClass(fxn_d "${my_instance}" "${bool_var_ptr}")
+        ct_assert_prints("a = bool_var, a dereferenced = TRUE")
     endfunction()
 endfunction()
 
