### 2023-08-14 82b6e4d
Remove underscore from permissible pointer specifiers
