### 2023-08-14 6b1f0f3
Add more tests for methods with pointer parameters
