+++ tests/class/class.cmake
@@ -169,6 +169,11 @@ function(${_test_member})
         function("${fxn_d}" self a)
             message("a = ${a}, a dereferenced = ${${a}}")
         endfunction()
+
+        cpp_member(fxn_e MyClass bool* bool)
+        function("${fxn_e}" self a b)
+            message("a = ${a}, a dereferenced = ${${a}}, b = ${b}")
+        endfunction()
     cpp_end_class()
 
     ct_add_section(NAME "_member_single_param")
@@ -199,6 +204,15 @@ function(${_test_member})
         set(bool_var_ptr bool_var)
         MyClass(fxn_d "${my_instance}" "${bool_var_ptr}")
         ct_assert_prints("a = bool_var, a dereferenced = TRUE")
+    endfunction()
+
+    ct_add_section(NAME "_member_bool_ptr_params")
+    function(${_member_bool_ptr_params})
+        MyClass(CTOR my_instance)
+        set(bool_var "TRUE")
+        set(bool_var_ptr bool_var)
+        MyClass(fxn_e "${my_instance}" "${bool_var_ptr}" "${bool_var}")
+        ct_assert_prints("a = bool_var, a dereferenced = TRUE, b = TRUE")
     endfunction()
 endfunction()
 
