### 2023-08-14 ff50383
Add virtual member tests for parameters and ptr params
