### 2023-08-18 bfc7171
Enable pointer type inference and add recursive pointer convertibility checking
