### 2023-08-21 #5711 @tianyi93
The document is about CMakePPLang, an object-oriented extension to the CMake language. It aims to simplify the creation and maintenance of projects built on CMake. The author has provided a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience. It also compared with CMake++ and explained the need for CMakePPLang.

Few comments are listed below:

- [ ] Do the authors clearly state what problems the software is designed to solve and who the target audience is?
The level of detail in the Statement of Need within the Documentation is currently not as comprehensive as that found in the paper. It is advisable to enhance the Statement of Need in the Documentation by incorporating the rationale behind the necessity for CMakePPLang. This will provide a clearer understanding of why CMakePPLang is required and how it contributes to the overall objectives.

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
The documentation provided example usage, however, did not state how to run automated tests.
