### 2023-08-24 6bdd056
Add convertibility tests for pointers
