+++ tests/class/class.cmake
@@ -46,9 +46,25 @@ function(${_test_class})
             ct_assert_equal(result TRUE)
         endfunction()
 
+        ct_add_section(NAME "_class_base_cast_ptr")
+        function(${_class_base_cast_ptr})
+            set(myclass_ptr "MyClass")
+            set(baseclass_ptr "BaseClass")
+            cpp_implicitly_convertible(result myclass_ptr baseclass_ptr)
+            ct_assert_equal(result TRUE)
+        endfunction()
+
         ct_add_section(NAME "_class_base_cast_obj")
         function(${_class_base_cast_obj})
             cpp_implicitly_convertible(result MyClass obj)
+            ct_assert_equal(result TRUE)
+        endfunction()
+
+        ct_add_section(NAME "_class_base_cast_obj")
+        function(${_class_base_cast_obj})
+            set(myclass_ptr MyClass)
+            set(obj_ptr obj)
+            cpp_implicitly_convertible(result myclass_ptr obj_ptr)
             ct_assert_equal(result TRUE)
         endfunction()
 
