### 2023-08-29 05f0d90
Fix broken class pointer cast tests
