### 2023-08-29 8e6eb55
Fix clobbered class cast to object test
