+++ tests/class/class.cmake
@@ -58,8 +58,8 @@ function(${_test_class})
             ct_assert_equal(result TRUE)
         endfunction()
 
-        ct_add_section(NAME "_class_base_cast_obj")
-        function(${_class_base_cast_obj})
+        ct_add_section(NAME "_class_base_cast_obj_ptr")
+        function(${_class_base_cast_obj_ptr})
             cpp_implicitly_convertible(result MyClass* obj*)
             ct_assert_equal(result TRUE)
         endfunction()
