### 2023-08-30 #5711 @bast
I am finally working on it - tracking my progress in this checklist:

* [ ]  **Repository:** Is the source code for this software available at the https://github.com/CMakePP/CMakePPLang/?
