### 2023-08-30 #5711 @bast
I still need to carefully go through my checklist above but for me the main question to address in the code documentation and in the paper is the question how the authors envision to manage versioning, compatibility, deprecation, and API changes in this tool. This tool is built on top of CMake language. The CMake language evolves (with its own versioning and deprectation mechanisms) and also CMakePPLang will evolve. How will it "move" with respect to an evolving CMake language underneath? If I as user had to decide whether to build my project on top of CMakePPLang, that would be a very important question for me. I would like to know what to expect 2-5 years down the line when CMake maybe has a new major version. Would my project stop working or would I have a way to pin a version? Do I need to anticipate lots of work updating my CMake code?
