### 2023-08-30 #5711 @bast
When reading the paper alone and looking at Figures 2 and 3 I was asking myself what "CTOR" meant. In the code documentation this is clarified but the figures would be easier to read if "CTOR" was explained at least in the code comment inside the example code ("GET" and "SET" are on the other hand intuitively clear).
