### 2023-08-30 #5711 @bast
A convincing argument to try this tool is its stronger typing. As a suggestion, it could be powerful to show in the paper a short example where because of weak typing a native CMake code could lead to an error that is difficult to catch whereas CMakePPLang catches the mistake with a clear error message since it is stricter about typing.
