### 2023-08-30 #5711 @bast
Version on https://cmakepp.github.io/CMakePPLang/ seems to be `1.0.0` but I am not sure this matches the version of the code. But maybe it is the target version?
