### 2023-09-01 91af2dc
Adjust figure size
