### 2023-09-01 795cca0
Adjust figure size
