### 2023-09-01 31a1b2c
Adjust figure size
