### 2023-09-01 7103978
Fix GIT_REPOSITORY link in 'Installation Instructions'
