### 2023-09-01 95c6dc9
Add section about version pinning in 'Obtaining CMakePPLang' section
