### 2023-09-01 1b1b9a7
Add discussion about current versioning scheme for releases to documentation and README
