### 2023-09-04 #5711 @tianyi93
> @tianyi93 when you get a chance please take a look at [CMakePP/CMakePPLang#113](https://github.com/CMakePP/CMakePPLang/pull/113). I believe it addresses your comments (and if it does not, please let me know).

LGTM, thanks!
