### 2023-09-05 #5711 @diehlpk
> > @tianyi93 when you get a chance please take a look at [CMakePP/CMakePPLang#113](https://github.com/CMakePP/CMakePPLang/pull/113). I believe it addresses your comments (and if it does not, please let me know).
>
>
>
> LGTM, thanks!

@tianyi93 can you please update the check list, if the author satisfied your comments.
