### 2023-09-05 e7fd626
Add pointer type covariance (#112)
* Enable pointer type inference and add recursive pointer convertibility checking

* Fix type mutation in find_ctor

* Add pointer tests for type_of

* Add convertibility tests for pointers

* Update documentation on pointer types

* Fix broken class pointer cast tests

* Fix clobbered class cast to object test

---------

Co-authored-by: Ryan Richard <ryanmrichard1@gmail.com>
