+++ tests/types/type_of.cmake
@@ -15,6 +15,13 @@ function("${test_cpp_type_of}")
         ct_assert_equal(return bool)
     endfunction()
 
+    ct_add_section(NAME "test_bool_ptr")
+    function("${test_bool_ptr}")
+        set(bool_ptr TRUE)
+        cpp_type_of(return bool_ptr)
+        ct_assert_equal(return bool*)
+    endfunction()
+
     ct_add_section(NAME "test_class")
     function("${test_class}")
         include(cmakepp_lang/class/class)
@@ -23,10 +30,26 @@ function("${test_cpp_type_of}")
         ct_assert_equal(return class)
     endfunction()
 
+    ct_add_section(NAME "test_class_ptr")
+    function("${test_class_ptr}")
+        include(cmakepp_lang/class/class)
+        cpp_class(a_class)
+        set(class_ptr a_class)
+        cpp_type_of(return class_ptr)
+        ct_assert_equal(return class*)
+    endfunction()
+
     ct_add_section(NAME "test_command")
     function("${test_command}")
         cpp_type_of(return add_subdirectory)
         ct_assert_equal(return fxn)
+    endfunction()
+
+    ct_add_section(NAME "test_command_ptr")
+    function("${test_command_ptr}")
+        set(command_ptr "add_subdirectory")
+        cpp_type_of(return command_ptr)
+        ct_assert_equal(return fxn*)
     endfunction()
 
     ct_add_section(NAME "test_desc")
@@ -35,10 +58,24 @@ function("${test_cpp_type_of}")
         ct_assert_equal(return desc)
     endfunction()
 
+    ct_add_section(NAME "test_desc_ptr")
+    function("${test_desc_ptr}")
+        set(desc_ptr "hello world")
+        cpp_type_of(return desc_ptr)
+        ct_assert_equal(return desc*)
+    endfunction()
+
     ct_add_section(NAME "empty_string")
     function("${empty_string}")
         cpp_type_of(return "")
         ct_assert_equal(return desc)
+    endfunction()
+
+    ct_add_section(NAME "empty_string_ptr")
+    function("${empty_string_ptr}")
+        set(str_ptr "")
+        cpp_type_of(return str_ptr)
+        ct_assert_equal(return desc*)
     endfunction()
 
     ct_add_section(NAME "test_float")
@@ -47,10 +84,24 @@ function("${test_cpp_type_of}")
         ct_assert_equal(return float)
     endfunction()
 
+    ct_add_section(NAME "test_float_ptr")
+    function("${test_float_ptr}")
+        set(float_ptr 3.14)
+        cpp_type_of(return float_ptr)
+        ct_assert_equal(return float*)
+    endfunction()
+
     ct_add_section(NAME "test_int")
     function("${test_int}")
         cpp_type_of(return 42)
         ct_assert_equal(return int)
+    endfunction()
+
+    ct_add_section(NAME "test_int_ptr")
+    function("${test_int_ptr}")
+        set(int_ptr 42)
+        cpp_type_of(return int_ptr)
+        ct_assert_equal(return int*)
     endfunction()
 
     ct_add_section(NAME "test_list")
@@ -59,12 +110,27 @@ function("${test_cpp_type_of}")
         ct_assert_equal(return list)
     endfunction()
 
+    ct_add_section(NAME "test_list_ptr")
+    function("${test_list_ptr}")
+        set(list_ptr hello world)
+        cpp_type_of(return list_ptr)
+        ct_assert_equal(return list*)
+    endfunction()
+
     ct_add_section(NAME "test_map")
     function("${test_map}")
         include(cmakepp_lang/map/map)
         cpp_map(CTOR a_map)
         cpp_type_of(result "${a_map}")
[…]
