+++ cmake/cmakepp_lang/class/find_ctor.cmake
@@ -62,7 +62,8 @@ function(_cpp_find_ctor _fc_this _fc_type)
         list(APPEND _fc_sig "desc")
         foreach(_fc_arg_i ${ARGN})
             cpp_type_of(_fc_type_i "${_fc_arg_i}")
-            cpp_sanitize_string(_fc_nice_type_i "${_fc_type_i}")
+            string(TOLOWER "${_fc_type_i}" _fc_nice_type_i)
+            #cpp_sanitize_string(_fc_nice_type_i "${_fc_type_i}")
             list(APPEND _fc_sig "${_fc_nice_type_i}")
         endforeach()
 
