### 2023-09-05 f9a1e75
Add strong typing advantage example
