### 2023-09-05 b781c9c
Add CMake list(LENGTH reference and remove snapshot paths
