+++ paper/paper.bib
@@ -13,7 +13,6 @@
 	month = may,
 	year = {2023},
 	pages = {184801},
-	file = {Snapshot:/mnt/2C627FE4627FB0E6/business/research/references/zotero/storage/QL3QKV8N/PluginPlay-Enabling-exascale-scientific-software.html:text/html},
 }
 
 @article{wirth_good_2006,
@@ -74,7 +73,14 @@
 	year = {2023},
 	language = {en-US},
 	urldate = {2023-06-26},
-	file = {Snapshot:/mnt/2C627FE4627FB0E6/business/research/references/zotero/storage/UL6Y5BBP/cmake.org.html:text/html},
+}
+
+@misc{cmake,
+	title = {{CMake list(LENGTH}},
+	url = {https://cmake.org/cmake/help/latest/command/list.html#length},
+	year = {2023},
+	language = {en-US},
+	urldate = {2023-09-05},
 }
 
 @misc{cmakepp_org,
@@ -85,7 +91,6 @@
 	language = {en-US},
 	urldate = {2023-06-26},
 	journal = {cmakepp.github.io},
-	file = {Snapshot:/mnt/2C627FE4627FB0E6/business/research/references/zotero/storage/BRWTNXJX/cmakepp.github.io.html:text/html},
 }
 
 @misc{cmaize,
