### 2023-09-05 83e1bad
Add CMake list(LENGTH reference and remove snapshot paths
