+++ paper/paper.bib
@@ -75,8 +75,8 @@
 	urldate = {2023-06-26},
 }
 
-@misc{cmake,
-	title = {{CMake list(LENGTH}},
+@misc{cmake_list_length,
+	title = {{CMake List Length}},
 	url = {https://cmake.org/cmake/help/latest/command/list.html#length},
 	year = {2023},
 	language = {en-US},
