### 2023-09-05 d4b02db
Add cpp_list(LENGTH signature to text and explain pointers
