### 2023-09-05 5318ac2
Add missing parentheses around CMakePPLang list length text reference
