+++ paper/paper.bib
@@ -83,6 +83,14 @@
 	urldate = {2023-09-05},
 }
 
+@misc{semver,
+	title = {{Semantic Versioning 2.0.0}},
+	url = {https://semver.org/},
+	year = {2023},
+	language = {en-US},
+	urldate = {2023-09-05},
+}
+
 @misc{cmakepp_org,
 	title = {{CMakePP} {Organization}},
 	url = {https://cmakepp.github.io/},
