### 2023-09-06 #5711 @bast
My comments have been addressed in https://github.com/CMakePP/CMakePPLang/pull/114 and https://github.com/CMakePP/CMakePPLang/pull/115 - thank you so much! After the changes are merged, can we build a new version of the manuscript for a final read-through? (If this is not the way how it usually works, I can also build it then locally)
