### 2023-09-06 9f406f4
Addressing comments from bast on the paper (#114)
* Clarify that 'CTOR' means 'constructor'

* Adjust figure size

* Adjust figure size

* Adjust figure size

* Add strong typing advantage example

* Add CMake list(LENGTH reference and remove snapshot paths

* Add CMake list(LENGTH reference and remove snapshot paths

* Add cpp_list(LENGTH signature to text and explain pointers

* Add missing parentheses around CMakePPLang list length text reference

* Add short statement about versioning
