### 2023-09-06 50eb54f
Addressing comments from bast on the repository (#115)
* Fix GIT_REPOSITORY link in 'Installation Instructions'

* Clarify that 'CTOR' means 'constructor' in Example Usage

* Add instructions for pinning specific version of CMakePPLang

* Add instructions for pinning specific version of CMakePPLang

* Add section about version pinning in 'Obtaining CMakePPLang' section

* Add discussion about current versioning scheme for releases to documentation and README

---------

Co-authored-by: Ryan Richard <ryanmrichard1@gmail.com>
