### 2023-09-12 #5711 @diehlpk
@ryanmrichard I read the paper, and the references are inconsistent. Sometimes there is a space `word (reference)` and sometimes there is no space `word(reference)`. I think both are fine, but it should be consistent.
