### 2023-09-12 #5711 @diehlpk
@ryanmrichard The C++ in the paper looks strange since the ++ are very big. One possible solution is to use `C\texttt{++}` to make it look nicer. This is not necessary, but I think the paper looks nicer.
