+++ paper/paper.md
@@ -36,36 +36,38 @@ bibliography: paper.bib
 
 # Summary
 
-CMakePPLang is an object-oriented extension to the CMake language written entirely using the original CMake language
-with the goal of making projects built on CMake easier to create and maintain. 
-That said, CMakePPLang has different coding practices, paradigms, and standards than the original CMake language, 
-much in the same way that C++ coding differs from C coding despite some level of interoperability. 
-Currently, CMakePPLang is used within the CMakePP organization[@cmakepp_org] as the 
-foundation for two in-progress projects: CMakeTest[@cmaketest] and CMaize[@cmaize]. CMakeTest
+CMakePPLang is an object-oriented extension to the CMake language written
+entirely using the original CMake language
+with the goal of making projects built on CMake easier to create and maintain.
+That said, CMakePPLang has different coding practices, paradigms, and standards than the original CMake language,
+much in the same way that C\texttt{++} coding differs from C coding despite some level of interoperability.
+Currently, CMakePPLang is used within the CMakePP organization [@cmakepp_org] as the
+foundation for two in-progress projects: CMakeTest [@cmaketest] and CMaize [@cmaize]. CMakeTest
 provides a solution for unit testing CMake and CMakePPLang code. CMaize
-is a CMake tool to simplify interoperability between projects and writing 
+is a CMake tool to simplify interoperability between projects and writing
 their build systems.
 
 
 # Statement of Need
 
-CMake[@cmake] is an extensible build tool that exceeds at generating build systems
+CMake [@cmake] is an extensible build tool that exceeds at generating build
+systems
 for many combinations of platforms, compilers, and build configurations.
-CMake has become the *de facto* standard tool for building C, C++, and
+CMake has become the *de facto* standard tool for building C, C\texttt{++}, and
 Fortran programs of moderate to large size. However, as the size of a project
 increases, the complexity of the CMake build code tends to increase as well,
 and the need arises to make building projects with CMake easier and less
 error prone. The complexity of builds will also increase as scientific
 computing moves toward heterogeneous systems, requiring programs to leverage
-a combination of CPUs, GPUs, and other specialized hardware. [@richard_pluginplay_2023] Better
+a combination of CPUs, GPUs, and other specialized hardware [@richard_pluginplay_2023].  Better
 utilities and extensions in CMake can help alleviate these issues, but these
-tools must be able to be designed in a maintainable and testable way.
-[@pressman_142_2005] Object-oriented programming excels at managing and
-maintaining large, complex code bases[@wirth_good_2006; @ambler_realistic_1998],
+tools must be able to be designed in a maintainable and testable way [@pressman_142_2005].
+ Object-oriented programming excels at managing and
+maintaining large, complex code bases [@wirth_good_2006; @ambler_realistic_1998],
 and there is an increasing need for this in the CMake language.
 
-Tobias Becker recognized these issues and wrote a purely object-oriented
-language on top of CMake, called CMake++ (formerly oo-cmake).[@becker_cmake_2021]
+Tobias Becker [@becker_cmake_2021] recognized these issues and wrote a purely object-oriented
+language on top of CMake, called CMake\texttt{++} (formerly oo-cmake) .
 CMake++ contains an abundance of CMake extensions. Many of those extensions
 have direct overlap with extensions that are part of CMakePPLang. Features include
 (among many): maps, objects, tasks/promises. Unfortunately development of
@@ -76,14 +78,14 @@ One of the primary issues with CMake++ is the lack of documentation. While
 there is some high-level documentation, there is little to no API or detailed
 developer documentation. This makes it very challenging for a new developer to figure out
 what is going on. Initially, forking and expanding on CMake++ was
[…]
