### 2023-09-12 09cf9e1
official tex solution for C++
