### 2023-09-12 #5711 @ryanmrichard
@diehlpk Thanks for pointing those issues out. I agree with both of your comments. I looked into the C++ thing a bit and apparently ISO C++ has their own [recommendation](https://isocpp.org/wiki/faq/misc-environmental-issues#latex-macros); I went with the ISO recommendation and think it looks nice.
