### 2023-09-12 #5711 @diehlpk
@ryanmrichard Thanks, now I am OK with the paper.

Please generate the following:

- [ ] A new release including all changes made during the JOSS review. Please post the new version number here.
- [ ] Upload that release to Zenodo or FigShare and post the DOI here. Note that authors and the title on the archive have to match with the paper.
