### 2023-09-12 #5711 @ryanmrichard
The version number of CMakePPLang is [v1.0.0](https://github.com/CMakePP/CMakePPLang/releases/tag/v1.0.0)

The Zenodo DOI is 10.5281/zenodo.8339016
