### 2023-09-12 #5711 @diehlpk
@ryanmrichard The title of the archive is wrong it should read as

CMakePPLang: An object-oriented extension to CMake

but it reads as

CMakePP/CMakePPLang: JOSS Submission
