### 2023-09-12 #5711 @diehlpk
@ryanmrichard The authors seem not to match as well.

In the paper, we have

`Zachery Crandall 1,2* , Blake Mulnix1,2* , Branden Butler1,2 , Theresa L. Windus 1,2 , and Ryan M. Richard`

and in the archive, we have

` Ryan Richard; Zachery Crandall; Blake Mulnix; Branden Butler`
