### 2023-09-12 d314498
add .zenodo.json file
