+++ CITATION.cff
@@ -1,19 +1,0 @@
-cff-version: 1.1.0
-message: "If you use this software, please cite it as below."
-authors:
-  - family-names: Crandall
-    given-names: Zachery
-    orcid: https://orcid.org/0000-0003-3161-9378
-  - family-names: Mulnix
-    given-names: Blake
-  - family-names: Butler
-    given-names: Branden
-  - family-names: Windus
-    given-names: Theresa L.
-    orcid: https://orcid.org/0000-0001-6065-3167
-  - family-names: Richard
-    given-names: Ryan M.
-    orcid: https://orcid.org/0000-0003-4235-5179
-title: CMakePPLang: An object-oriented extension to CMake
-version: v1.0.0
-date-released: 2023-09-12
