### 2023-09-12 #5711 @ryanmrichard
@diehlpk I think I fixed it the new DOI is 10.5281/zenodo.8339121 and it's version v1.0.3
