### 2023-09-13 #5711 @danielskatz
@ryanmrichard - see https://github.com/CMakePP/CMakePPLang/pull/119 for some small changes in the paper. Please merge this or let me know what you disagree with, then we can continue the process
