### 2023-09-13 e121a13
minor changes in paper and bib (#119)
* minor changes in paper

* minor bib changes
