+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{richard_pluginplay_2023,
-	title = {{PluginPlay}: {Enabling} exascale scientific software one module at a time},
+	title = {{PluginPlay}: Enabling exascale scientific software one module at a time},
 	volume = {158},
 	issn = {0021-9606},
 	shorttitle = {{PluginPlay}},
@@ -34,10 +34,10 @@
 
 @incollection{pressman_142_2005,
 	edition = {7},
-	title = {{Software} {Quality}},
+	title = {Software Quality},
 	isbn = {978-0-07-301933-8},
 	language = {en},
-	booktitle = {Software {Engineering}: {A} {Practitioner}'s {Approach}},
+	booktitle = {Software Engineering: A Practitioner's Approach},
 	publisher = {Palgrave Macmillan},
 	collaborator = {Pressman, Roger S.},
 	year = {2005},
@@ -45,7 +45,7 @@
 }
 
 @misc{ambler_realistic_1998,
-	title = {A {Realistic} {Look} at {Object}-{Oriented} {Reuse}},
+	title = {A Realistic Look at Object-Oriented Reuse},
 	url = {http://www.drdobbs.com/a-realistic-look-at-object-oriented-reus/184415594},
 	abstract = {Software tools and techniques for global software development. Dr. Dobb's features articles, source code, blogs,forums,video tutorials, and audio podcasts, as well as articles from Dr. Dobb's Journal, BYTE.com, C/C++ Users Journal, and Software Development magazine.},
 	urldate = {2023-06-12},
@@ -76,7 +76,7 @@
 }
 
 @misc{cmake_list_length,
-	title = {{CMake List Length}},
+	title = {{CMake} List Length},
 	url = {https://cmake.org/cmake/help/latest/command/list.html#length},
 	year = {2023},
 	language = {en-US},
@@ -84,7 +84,7 @@
 }
 
 @misc{semver,
-	title = {{Semantic Versioning 2.0.0}},
+	title = {Semantic Versioning 2.0.0},
 	url = {https://semver.org/},
 	year = {2023},
 	language = {en-US},
@@ -92,7 +92,7 @@
 }
 
 @misc{cmakepp_org,
-	title = {{CMakePP} {Organization}},
+	title = {{CMakePP} Organization},
 	url = {https://cmakepp.github.io/},
 	abstract = {Website for CMake Packaging Project},
 	year = {2023},
@@ -137,7 +137,7 @@
 }
 
 @misc{python_map_2023,
-	title = {Built-in {Types}: {Mapping} {Types} - dict},
+	title = {Built-in {T}ypes: Mapping {T}ypes - dict},
 	url = {https://docs.python.org/3/library/stdtypes.html#mapping-types-dict},
 	abstract = {A mapping object maps hashable values to arbitrary objects. Mappings are mutable objects. There is currently only one standard mapping type, the dictionary. (For other containers see the built-in list, set, and tuple classes, and the collections module.)},
 	urldate = {2023-06-12},
