### 2023-09-14 #5711 @danielskatz
Congratulations to @ryanmrichard (Ryan M. Richard) and co-authors on your publication!!

JOSS depends on volunteers like you
